+++ packages/featureserver/src/route.js
@@ -1,3 +1,4 @@
+const geojsonhint = require('@mapbox/geojsonhint')
 const FsInfo = require('./info.js')
 const FsQuery = require('./query.js')
 const FsGenerateRenderer = require('./generateRenderer')
@@ -14,6 +15,12 @@ module.exports = route
  * @param  {Object}   geojson
  */
 function route (req, res, geojson, options) {
+  // Check for valid GeoJSON and warn if not found (non-production environments)
+  if (process.env.NODE_ENV !== 'production') {
+    let geojsonErrors = geojsonhint.hint(geojson)
+    if (geojsonErrors.length > 0) console.log(`\nWARNING: Source data for ${req.path} is invalid GeoJSON:\n ${geojsonErrors.map((err, i) => `\t ${i + 1}) ${err.message}\n`)}`)
+  }
+
   req.query = req.query || {}
   req.query = coerceQuery(req.query)
   req.params = req.params || {}